Rename FitResult to FitResultBase, add reset hooks
