+++ src/easydiffraction/analysis/__init__.py
@@ -11,7 +11,7 @@ from easydiffraction.analysis.categories.fit_parameter_correlations import (
 from easydiffraction.analysis.categories.fit_parameters import FitParameterItem
 from easydiffraction.analysis.categories.fit_parameters import FitParameters
 from easydiffraction.analysis.categories.fit_parameters import FitParametersFactory
-from easydiffraction.analysis.categories.fit_result import FitResult
+from easydiffraction.analysis.categories.fit_result import FitResultBase
 from easydiffraction.analysis.categories.fit_result import FitResultFactory
 from easydiffraction.analysis.categories.joint_fit import JointFitCollection
 from easydiffraction.analysis.categories.joint_fit import JointFitFactory
