+++ src/easydiffraction/analysis/categories/__init__.py
@@ -11,7 +11,7 @@ from easydiffraction.analysis.categories.fit_parameter_correlations import (
 from easydiffraction.analysis.categories.fit_parameter_correlations import FitParameterCorrelations
 from easydiffraction.analysis.categories.fit_parameters import FitParameterItem
 from easydiffraction.analysis.categories.fit_parameters import FitParameters
-from easydiffraction.analysis.categories.fit_result import FitResult
+from easydiffraction.analysis.categories.fit_result import FitResultBase
 from easydiffraction.analysis.categories.fitting_mode import FittingMode
 from easydiffraction.analysis.categories.fitting_mode import FittingModeFactory
 from easydiffraction.analysis.categories.joint_fit import JointFitCollection
