+++ src/easydiffraction/analysis/categories/fit_result/__init__.py
@@ -1,5 +1,5 @@
 # SPDX-FileCopyrightText: 2026 EasyScience contributors <https://github.com/easyscience>
 # SPDX-License-Identifier: BSD-3-Clause
 
-from easydiffraction.analysis.categories.fit_result.default import FitResult
+from easydiffraction.analysis.categories.fit_result.base import FitResultBase
 from easydiffraction.analysis.categories.fit_result.factory import FitResultFactory
